Moved derp server code under pkg directory
Nexodus doesn't touch derp server code, so pkg directory
is the better place for the code.

The only remaining code that we have in the internal/nexodus
now is related to the integration with nexodus agent.
