+++ pkg/derp/testdata/example_ss.txt
